refactor(server): move entrypoints into src
