+++ server/package.json
@@ -2,7 +2,7 @@
   "name": "server",
   "version": "1.0.0",
   "description": "Discord Clone Backend",
-  "main": "index.js",
+  "main": "src/index.js",
   "type": "module",
   "scripts": {
     "start": "node src/index.js",
